+++ ConcordiumWallet/Model/CIS2/CIS2Token.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct CIS2Token: Codable {
+    let id: Int
+    let token, totalSupply: String
+}
