+++ ConcordiumWallet/Model/CIS2/CIS2Tokens.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct CIS2Tokens: Codable {
+    let count: Int
+    let limit: Int
+    let tokens: [CIS2Token]
+}
