+++ ConcordiumWallet/Model/CIS2/CIS2TokensMetadata.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct CIS2TokensMetadata: Codable {
+    var contractName: String
+    var metadata: [CIS2TokensMetadataItem]
+}
