+++ ConcordiumWallet/Model/CIS2/CIS2TokensMetadataItem.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct CIS2TokensMetadataItem: Codable {
+    var metadataChecksum: String?
+    var metadataURL: String
+    var tokenId: String
+}
